ai-engine/app/extensions.py
from flask_cors import CORS
from flask_sqlalchemy import SQLAlchemy
from flask_migrate import Migrate
from flask_jwt_extended import JWTManager
from flask_mail import Mail
from redis import Redis

cors = CORS()
db = SQLAlchemy()
migrate = Migrate()
jwt = JWTManager()
mail = Mail()
redis_client = Redis()
