ai-engine/requirements.txt
Flask>=3.0.0
flask-cors
flask-jwt-extended
flask-mail
gunicorn
python-dotenv
pydantic
email-validator
langchain
openai
celery
redis
elasticsearch
psycopg2-binary
minio
chromadb
